pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skillcheck"
version = "0.1.0"
description = "Validates Claude Code SKILL.md files against the Anthropic skill specification"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
tiktoken = [
    "tiktoken>=0.7",
]

[project.scripts]
skillcheck = "skillcheck.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/skillcheck"]

[tool.pytest.ini_options]
testpaths = ["tests"]
